added logic to control the scrol of the list view

+++ lib/src/controller.dart
@@ -45,7 +45,7 @@ class MultiDirectionalHorizontalListController {
   /// Command the list to switch to the given [position] in a synchronous and
   /// immediate manner. There will be no animation. To animate, use
   /// [animateToPosition] instead.
-  void jumpToPosition(int position) {
+  void jumpToPosition(double position) {
     _feedback?.sink.add(ControllerFeedback(
       ControllerCommandTypes.jumpToPosition,
       data: position,
@@ -54,7 +54,7 @@ class MultiDirectionalHorizontalListController {
 
   /// Command the list to move to the given [position] in an animated manner.
   /// To ignore animation, use [jumpToPosition] instead
-  Future<void> animateToPosition(int position) async {
+  Future<void> animateToPosition(double position) async {
     _feedback?.sink.add(ControllerFeedback(
       ControllerCommandTypes.animateToPosition,
       data: position,
